refactor: enum types
