+++ src/types/enums/depth-index.ts
@@ -1,0 +1,19 @@
+/**
+ * Enumeration representing different depth levels.
+ *
+ * @enum {number}
+ * @property {number} L1 - Level 1 depth.
+ * @property {number} L2 - Level 2 depth.
+ * @property {number} L3 - Level 3 depth.
+ * @property {number} L4 - Level 4 depth.
+ * @property {number} L5 - Level 5 depth.
+ * @property {number} L6 - Level 6 depth.
+ */
+export enum DepthIndex {
+  L1 = 1,
+  L2 = 2,
+  L3 = 3,
+  L4 = 4,
+  L5 = 5,
+  L6 = 6
+}
