+++ src/types/enums/ecosystem.ts
@@ -1,0 +1,9 @@
+export enum Ecosystem {
+  EVM = 'EVM',
+  Bitcoin = 'Bitcoin',
+  Solana = 'Solana',
+  Tron = 'Tron',
+  Ton = 'Ton',
+  IC = 'ICP',
+  Kaspa = 'Kaspa'
+}
