+++ src/types/enums/index.ts
@@ -1,0 +1,13 @@
+export { ChainAssetType } from './chain-asset-type';
+export { DepthIndex } from './depth-index';
+export { Ecosystem } from './ecosystem';
+export { SwapV2ExactTokensType } from './exact-token-type';
+export { OrderCategory } from './order-category';
+export { OrderSide } from './order-side';
+export { TickSpace } from './tick-space';
+export { TransactionType } from './tx-type';
+export { WalletSignatureSchema } from './wallet-sign-schema';
+
+export enum Version {
+  V0 = 0
+}
