+++ src/types/enums/order-category.ts
@@ -1,0 +1,11 @@
+/**
+ * Enum representing the different categories of orders.
+ *
+ * @enum {number}
+ * @property {number} LimitOrder - A limit order.
+ * @property {number} SwapOrder - A swap order.
+ */
+export enum OrderCategory {
+  LimitOrder,
+  SwapOrder
+}
