+++ src/types/enums/order-side.ts
@@ -1,0 +1,14 @@
+/**
+ * Enum representing the side of an order.
+ */
+export enum OrderSide {
+  /**
+   * Ask
+   */
+  Ask = 1,
+
+  /**
+   * Bid
+   */
+  Bid = 2
+}
