+++ src/types/enums/tick-space.ts
@@ -1,0 +1,49 @@
+/**
+ * Enum representing different time intervals for tick spacing.
+ */
+export enum TickSpace {
+  /**
+   * 1 minute
+   */
+  OneMinute = 1,
+
+  /**
+   * 5 minutes
+   */
+  FiveMinutes = 2,
+
+  /**
+   * 15 minutes
+   */
+  FifteenMinutes = 3,
+
+  /**
+   * 30 minutes
+   */
+  ThirtyMinutes = 4,
+
+  /**
+   * 1 hour
+   */
+  OneHour = 5,
+
+  /**
+   * 4 hours
+   */
+  FourHours = 6,
+
+  /**
+   * 1 day
+   */
+  OneDay = 7,
+
+  /**
+   * 1 week
+   */
+  OneWeek = 8,
+
+  /**
+   * 1 month
+   */
+  OneMonth = 9
+}
