+++ src/types/enums/wallet-sign-schema.ts
@@ -1,0 +1,9 @@
+export enum WalletSignatureSchema {
+  BtcEcdsa = '0x3e8',
+  EvmEcdsa = '0x3e9',
+  TronEcdsa = '0x3ea',
+  TonEddsaOpenMask = '0x7d0',
+  SolanaEddsa = '0x7da',
+  IcpEddsa = '0x7e4',
+  KaspaSchnorr = '0xbc2'
+}
